+++ src/promptflow/promptflow/_sdk/entities/_trace.py
@@ -12,13 +12,15 @@ from google.protobuf.json_format import MessageToJson
 from opentelemetry.proto.trace.v1.trace_pb2 import Span as PBSpan
 
 from promptflow._constants import (
-    DEFAULT_SESSION_ID,
     DEFAULT_SPAN_TYPE,
     SpanAttributeFieldName,
     SpanContextFieldName,
     SpanFieldName,
+    SpanResourceAttributesFieldName,
+    SpanResourceFieldName,
     SpanStatusFieldName,
 )
+from promptflow._sdk._constants import CumulativeTokenCountFieldName
 from promptflow._sdk._orm.trace import Span as ORMSpan
 from promptflow._sdk._utils import (
     convert_time_unix_nano_to_timestamp,
@@ -116,7 +118,7 @@ class Span:
 
     @staticmethod
     def _from_protobuf_object(obj: PBSpan, resource: typing.Dict) -> "Span":
-        span_dict = json.loads(MessageToJson(obj))
+        span_dict: dict = json.loads(MessageToJson(obj))
         span_id = obj.span_id.hex()
         trace_id = obj.trace_id.hex()
         context = {
@@ -130,13 +132,17 @@ class Span:
         status = {
             SpanStatusFieldName.STATUS_CODE: parse_otel_span_status_code(obj.status.code),
         }
-        attributes = flatten_pb_attributes(span_dict[SpanFieldName.ATTRIBUTES])
-        # `session_id` and `span_type` are not standard fields in OpenTelemetry attributes
+        # we have observed in some scenarios, there is not `attributes` field
+        attributes = flatten_pb_attributes(span_dict.get(SpanFieldName.ATTRIBUTES, dict()))
+        # `span_type` are not standard fields in OpenTelemetry attributes
         # for example, LangChain instrumentation, as we do not inject this;
-        # so we need to get them with default value to avoid KeyError
+        # so we need to get it with default value to avoid KeyError
         span_type = attributes.get(SpanAttributeFieldName.SPAN_TYPE, DEFAULT_SPAN_TYPE)
-        # note that this might make these spans persisted in another partion if we split the trace table by `session_id`
-        session_id = attributes.get(SpanAttributeFieldName.SESSION_ID, DEFAULT_SESSION_ID)
+
+        # parse from resource.attributes: session id, experiment
+        resource_attributes: dict = resource[SpanResourceFieldName.ATTRIBUTES]
+        session_id = resource_attributes[SpanResourceAttributesFieldName.SESSION_ID]
+        experiment = resource_attributes.get(SpanResourceAttributesFieldName.EXPERIMENT_NAME, None)
 
         return Span(
             name=obj.name,
@@ -150,6 +156,7 @@ class Span:
             span_type=span_type,
             session_id=session_id,
             parent_span_id=parent_span_id,
+            experiment=experiment,
         )
 
 
@@ -162,11 +169,11 @@ class _LineRunData:
     root_span_id: str
     inputs: typing.Dict
     outputs: typing.Dict
-    start_time: datetime.datetime
-    end_time: datetime.datetime
+    start_time: str
+    end_time: str
     status: str
     latency: float
-    name: str
+    display_name: str
     kind: str
     cumulative_token_count: typing.Optional[typing.Dict[str, int]]
 
@@ -188,9 +195,9 @@ class _LineRunData:
         # if there is no token usage, set `cumulative_token_count` to None
         if total_token_count > 0:
             cumulative_token_count = {
-                "completion": completion_token_count,
-                "prompt": prompt_token_count,
-                "total": total_token_count,
+                CumulativeTokenCountFieldName.COMPLETION: completion_token_count,
+                CumulativeTokenCountFieldName.PROMPT: prompt_token_count,
+                CumulativeTokenCountFieldName.TOTAL: total_token_count,
             }
         else:
             cumulative_token_count = None
@@ -205,7 +212,7 @@ class _LineRunData:
             end_time=end_time.isoformat(),
             status=span._content[SpanFieldName.STATUS][SpanStatusFieldName.STATUS_CODE],
             latency=(end_time - start_time).total_seconds(),
-            name=span.name,
+            display_name=span.name,
             kind=attributes.get(SpanAttributeFieldName.SPAN_TYPE, span.span_type),
             cumulative_token_count=cumulative_token_count,
         )
@@ -224,29 +231,35 @@ class LineRun:
     end_time: str
     status: str
     latency: float
-    name: str
+    display_name: str
     kind: str
     cumulative_token_count: typing.Optional[typing.Dict[str, int]] = None
     evaluations: typing.Optional[typing.List[typing.Dict]] = None
 
     @staticmethod
-    def _from_spans(spans: typing.List[Span]) -> "LineRun":
+    def _from_spans(spans: typing.List[Span]) -> typing.Optional["LineRun"]:
         main_line_run_data: _LineRunData = None
-        evaluation_line_run_data_dict = dict()
+        evaluations = []
         for span in spans:
             if span.parent_span_id:
                 continue
             attributes = span._content[SpanFieldName.ATTRIBUTES]
-            if SpanAttributeFieldName.REFERENCED_LINE_RUN_ID in attributes:
-                evaluation_line_run_data_dict[span.name] = _LineRunData._from_root_span(span)
+            if (
+                SpanAttributeFieldName.REFERENCED_LINE_RUN_ID in attributes  # test scenario
+                or SpanAttributeFieldName.REFERENCED_BATCH_RUN_ID in attributes  # batch run scenario
+            ):
+                evaluations.append(_LineRunData._from_root_span(span))
             elif SpanAttributeFieldName.LINE_RUN_ID in attributes:
                 main_line_run_data = _LineRunData._from_root_span(span)
             else:
                 # eager flow/arbitrary script
                 main_line_run_data = _LineRunData._from_root_span(span)
-        evaluations = dict()
-        for eval_name, eval_line_run_data in evaluation_line_run_data_dict.items():
-            evaluations[eval_name] = eval_line_run_data
+        # main line run span is absent, ignore this line run
+        # this may happen when the line is still executing, or terminated;
+        # or the line run is killed before the traces exported
+        if main_line_run_data is None:
+            return None
+
         return LineRun(
             line_run_id=main_line_run_data.line_run_id,
             trace_id=main_line_run_data.trace_id,
@@ -257,7 +270,7 @@ class LineRun:
             end_time=main_line_run_data.end_time,
             status=main_line_run_data.status,
             latency=main_line_run_data.latency,
-            name=main_line_run_data.name,
+            display_name=main_line_run_data.display_name,
             kind=main_line_run_data.kind,
             cumulative_token_count=main_line_run_data.cumulative_token_count,
             evaluations=evaluations,